// ClientApp/src/app/model/migration-survey.model.ts
export interface MigrationSurvey {
  id: number;
  district: string;
  block: string;
  gramPanchayat: string;
  revenueVillage: string;
  totalHouseholds: number;
  malePopulation: number;
  femalePopulation: number;
  totalPopulation: number;
  internalVillageRoads: boolean | null;
  internalVillageRoadsRequirement: string | null;
  internalDrainsAvailable: boolean | null;
  drainsProperlyFunctional: boolean | null;
  isElectrified: boolean | null;
  streetLightingAvailable: boolean | null;
  streetLightingType: string | null;
  villageConnectedToGP: string | null;
  lengthAllWeatherRoadToGP: number | null;
  gpConnectedToPWDOrHighway: boolean | null;
  lengthAllWeatherRoadToHighway: number | null;
  menInMigration: number | null;
  womenInMigration: number | null;
  minorChildrenInMigration: number | null;
  totalPersonsInMigration: number;
  drinkingWaterSource: string | null;
  allHouseholdsWithToilets: boolean | null;
  anganwadiCentre: boolean | null;
  primarySchoolAvailable: boolean | null;
  secondarySchoolWithin3km: boolean | null;
  subHealthCentre: boolean | null;
  communityCentreAvailable: boolean | null;
  commonShedForWSHG: boolean | null;
  playgroundAvailable: boolean | null;
  communityTanks: number | null;
  mobileNetworkCoverage: boolean | null;
  digitalConnectivity: boolean | null;
  dryingYard: boolean | null;
  pdsAvailable: boolean | null;
  distanceOfPDS: number | null;
  bankingPostOfficeNearby: boolean | null;
  waterFromIrrigationProject: boolean | null;
  repairOrNewDistributionCanalRequired: boolean | null;
  lengthOfDistributionCanal: number | null;
  functionalLiftIrrigation: boolean | null;
  scopeOfNewLiftIrrigation: boolean | null;
  functionalCheckDams: boolean | null;
  scopeOfNewCheckDams: boolean | null;
  functionalDistributionCanal: boolean | null;
  scopeOfNewDistributionCanal: number | null;
  respondentName: string | null;
  identityRole: string | null;
  surveyProcess: string | null;
  respondentMobile: string | null;
  meetingPhotoPath: string | null;
  geoLocation: string | null;
  enumeratorName: string | null;
  surveyDate: string;
}